requirements.txt
flask==2.3.2
torch==2.2.0
torchvision==0.17.0
opencv-python==4.10.0.84
numpy==1.26.0
pillow==10.0.0
SQLAlchemy==2.0.44
werkzeug==2.3.7
gunicorn==21.2.0
